:white_check_mark: Refactor connectors testing
