+++ pyproject.toml
@@ -48,7 +48,8 @@ dynamic = ["version"]
 dev = ["tox", "pre-commit", "black", "flake8", "isort", "mypy", "pytest", "coverage", "build", "ruff"]
 test = ["pytest", "coverage", "pytest-cov", "google-auth-oauthlib", "google-api-python-client"]
 transformers = ["transformers>4"]
-connectors = ["exchangelib", "google-auth-oauthlib", "google-api-python-client"]
+connectors = ["exchangelib"]
+gmail = ["google-auth-oauthlib", "google-api-python-client"]
 docs = ["mkdocs", "markdown", "mkdocs-material", "mdx-include"]
 
 [tool.setuptools.packages.find]
