+++ tests/conftest.py
@@ -11,7 +11,6 @@ from melusine import config
 pytest_plugins = [
     "tests.fixtures.backend",
     "tests.fixtures.basic_emails",
-    "tests.fixtures.connectors",
     "tests.fixtures.docs",
     "tests.fixtures.pipelines",
     "tests.fixtures.processors",
