+++ tests/gmail/test_gmail.py
@@ -1,14 +1,62 @@
-import base64
 import logging
 import os
-from unittest.mock import MagicMock, patch
+import pytest
 
 import pandas as pd
-import pytest
-from google.oauth2.credentials import Credentials
-from googleapiclient.http import HttpRequestMock
+from unittest.mock import MagicMock, patch
 
+
+HttpRequestMock = pytest.importorskip('googleapiclient.http.HttpRequestMock')
+from google.oauth2.credentials import Credentials
 from melusine.connectors.gmail import GmailConnector
+
+
+def return_value(resp, content):
+    return content
+
+
+@pytest.fixture
+def mocked_gc():
+    with patch("melusine.connectors.gmail.build") as mock_build:
+        with patch("melusine.connectors.gmail.Credentials.from_authorized_user_file") as mock_creds_from_file:
+            with patch("melusine.connectors.gmail.os.path.exists") as mock_exists:
+                mock_exists.return_value = True
+                mock_service = MagicMock()
+                mock_service.users().getProfile.return_value = HttpRequestMock(
+                    None, {"emailAddress": "test@example.com"}, return_value
+                )
+                mock_service.users().labels().list.return_value = HttpRequestMock(
+                    None,
+                    {
+                        "labels": [
+                            {"id": "INBOX", "name": "INBOX", "type": "system"},
+                            {
+                                "id": "TRASH",
+                                "name": "TRASH",
+                                "messageListVisibility": "hide",
+                                "labelListVisibility": "labelHide",
+                                "type": "system",
+                            },
+                            {"id": "UNREAD", "name": "UNREAD", "type": "system"},
+                        ]
+                    },
+                    return_value,
+                )
+                mock_build.return_value = mock_service
+                mock_creds_from_file.return_value = Credentials("dummy")
+
+                return GmailConnector(token_json_path="token.json", done_label="TRASH", target_column="target")
+
+
+@pytest.fixture
+def fake_image():
+    image_data = b""
+    width = height = 100
+    for _ in range(height):
+        row_data = b"\xff" * (width * 3)
+        image_data += row_data
+
+    return image_data
 
 
 def return_value(resp, content):
@@ -266,4 +314,4 @@ def test_gc_send_email(mocked_gc, fake_image, caplog):
             {"attachment.jpg": fake_image},
         )
 
-    assert "Email sent to melusine_testing@yopmail.com, Message Id: 12456"
+    assert "Email sent to melusine_testing@yopmail.com, Message Id: 12456"  in caplog.text
