+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
 requires =
     tox>=4
-env_list = clean, core38, core310, transformers, report
+env_list = clean, core38, core310, transformers, gmail, report
 
 [gh-actions]
 python =
-  3.8: clean, core38, transformers
+  3.8: clean, core38, transformers, gmail
   3.10: core310
 
 [testenv]
@@ -13,8 +13,6 @@ commands = pytest --cov --cov-append --cov-report xml
 deps =
     pytest
     pytest-cov
-    google-auth-oauthlib
-    google-api-python-client
 depends =
     {core38,transformers}: clean
     report: core38,transformers
@@ -37,6 +35,12 @@ description = run unit tests with the transformers dependency
 deps={[testenv]deps}
 commands = pytest tests/huggingface --cov --cov-append --cov-report xml
 extras = transformers
+
+[testenv:gmail]
+description = run unit tests with the gmail dependencies
+deps={[testenv]deps}
+commands = pytest tests/gmail --cov --cov-append --cov-report xml
+extras = gmail
 
 [testenv:report]
 deps = coverage[toml]
